refactor: メンバー機能をfeatureディレクトリに集約
