+++ app/web/app/.server/actions/index.ts
@@ -1,3 +1,2 @@
 export * from "./friendAction";
-export * from "./memberAction";
 export * from "./paymentAction";
