+++ app/web/app/.server/loaders/index.ts
@@ -1,3 +1,2 @@
 export * from "./friendsLoader";
-export * from "./membersLoader";
 export * from "./paymentsLoader";
