+++ app/web/app/features/members/components/add-member-form.tsx
@@ -1,11 +1,13 @@
 import { Form, useLocation, useNavigation } from "@remix-run/react";
 import { useId } from "react";
-import { MemberAddList } from "~/components/MemberAddList";
+
 import { Button } from "~/components/ui/button";
 import { Input } from "~/components/ui/input";
 import { Label } from "~/components/ui/label";
 
-export function MemberAddForm() {
+import { AddMemberList } from "./add-member-list";
+
+export function AddMemberForm() {
 	const { search } = useLocation();
 	const { state } = useNavigation();
 	const searchParams = new URLSearchParams(search);
@@ -36,7 +38,7 @@ export function MemberAddForm() {
 				</div>
 			</Form>
 			<div className="flex flex-col gap-3 w-full h-80 overflow-y-auto">
-				<MemberAddList />
+				<AddMemberList />
 			</div>
 		</div>
 	);
