+++ app/web/app/features/members/components/index.ts
@@ -1,0 +1,5 @@
+export * from "./add-member-card";
+export * from "./add-member-form";
+export * from "./add-member-list";
+export * from "./member-card";
+export * from "./member-list";
