+++ app/web/app/features/members/components/member-card.tsx
@@ -1,4 +1,5 @@
 import type { User } from "~/api/@types";
+
 import { Avatar, AvatarFallback, AvatarImage } from "~/components/ui/avatar";
 
 interface Props {
