+++ app/web/app/features/members/loaders/index.ts
@@ -1,0 +1,1 @@
+export * from "./member-list-loader";
